refactor(voice): extract tuned prompt factories into voice-agent-config.ts (step 5a-1)
PURE MOVE of the greeting factory and system-instruction factory out of
voice-agent.ts (which runs main() at module load and is therefore
untestable) into an importable module. voice-agent.ts keeps its state
(mode resolver, meeting flag, session gates) and threads it in through
VoiceConfigContext; env-dependent reads (stand identity, voice context,
repo URL, agent context) run verbatim by default with test-only overrides.

Equivalence evidence:
- all 87 tuned string-array entries identical IN ORDER between the old
  region and the new module (programmatic comparison)
- template-literal set identical except the intentional override branch
- behavior anchors upgraded from source-hash tripwires to REAL factory
  output: fixed-env instruction hash, exact meeting-greeting string,
  conditional-variant assertions (meeting/googleSearch/mode-marker
  placement), reconnect replay-guard shape — 10/10 green across normal,
  repeat, and empty-CLAUDE_CONFIG_DIR runs

tsc clean; task-bridge suites green. Phone (conversation-server) untouched.
