+++ tests/fixtures/voice-behavior-anchors.json
@@ -1,5 +1,5 @@
 {
- "note": "Step-5 behavior anchors. A diff here means the tuned prompt surface changed — deliberate changes must regenerate via ANCHOR_UPDATE=1 and call the diff out in the PR.",
+ "note": "Step-5 behavior anchors (post-5a-1: real factory output). Deliberate prompt changes must regenerate via ANCHOR_UPDATE=1 with the diff called out in the PR.",
  "tools": [
   {
    "name": "activate_screen_companion",
@@ -1076,9 +1076,9 @@
    }
   }
  ],
+ "instructions_hash_fixed_env": "1fc9401c9d5f1845435781eb81ac92d1766e8902993268de5d972368ae81daaf",
+ "meeting_greeting": "[System: MEETING MODE — LISTEN AND TAKE NOTES. A Zoom meeting is active. Listen to everything and mentally track the discussion: who said what, key decisions, action items, topics covered. But do NOT produce any audio output UNLESS someone says \"Sutando\" or \"hey Sutando\" — then respond to their request using your accumulated notes and context. When not addressed, produce absolutely zero words — no acknowledgments, no \"silent\", no sounds. You are an invisible note-taker until called upon.]\n\n[MODE-MARKER-MEETING]",
  "regions": {
-  "instructions_factory": "f00f0ea80c15bda9150e42c3e8a94f4f4894991db21b52c0d19ff1f0525eca58",
-  "greeting_factory": "a3a67a3b628c831e98be2c82c874b17048eb920f901d14d4c288c5e241bdda16",
   "tool_table": "e6d53e7a98b6c0f052148b6bae6096eab4e5768fb8187be8d03deee179023e78"
  }
 }
